requirements.txt
watchdog
extract-msg
pydantic-ai
streamlit
logfire